Refactor the service parser
